+++ src/components/TextField/index.js
@@ -151,7 +151,10 @@ TextField.propTypes = {
     'tel',
   ]).isRequired,
   /** Text within the input */
-  value: PropTypes.string,
+  value: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.number,
+  ]),
 };
 
 export default TextField;